+++ src/main/java/com/vidcentral/api/application/member/MemberMapper.java
@@ -1,8 +1,8 @@
 package com.vidcentral.api.application.member;
 
 import com.vidcentral.api.domain.member.entity.Member;
-import com.vidcentral.api.dto.request.SignUpRequest;
-import com.vidcentral.api.dto.response.MemberInfoResponse;
+import com.vidcentral.api.dto.request.member.SignUpRequest;
+import com.vidcentral.api.dto.response.member.MemberInfoResponse;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
