+++ src/main/java/com/vidcentral/api/application/member/MemberService.java
@@ -3,17 +3,16 @@ package com.vidcentral.api.application.member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.vidcentral.api.application.auth.AuthenticationMapper;
 import com.vidcentral.api.application.auth.AuthorizationService;
 import com.vidcentral.api.domain.auth.entity.AuthMember;
 import com.vidcentral.api.domain.member.entity.Member;
-import com.vidcentral.api.dto.request.LoginRequest;
-import com.vidcentral.api.dto.request.SignUpRequest;
-import com.vidcentral.api.dto.request.TokenRequest;
-import com.vidcentral.api.dto.request.UpdateMemberRequest;
-import com.vidcentral.api.dto.response.LoginResponse;
-import com.vidcentral.api.dto.response.MemberInfoResponse;
+import com.vidcentral.api.dto.request.auth.LoginRequest;
+import com.vidcentral.api.dto.request.member.SignUpRequest;
+import com.vidcentral.api.dto.request.member.UpdateMemberRequest;
+import com.vidcentral.api.dto.response.auth.LoginResponse;
+import com.vidcentral.api.dto.response.member.MemberInfoResponse;
 
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -34,15 +33,12 @@ public class MemberService {
 	}
 
 	@Transactional
-	public LoginResponse loginMember(LoginRequest loginRequest) {
+	public LoginResponse loginMember(HttpServletResponse httpServletResponse, LoginRequest loginRequest) {
 		final Member loginMember = memberReadService.findMember(loginRequest.email());
 		memberReadService.validateLoginPasswordMatch(loginRequest.password(), loginMember.getPassword());
 
-		TokenRequest tokenRequest = authorizationService
-			.issueServiceToken(loginMember.getEmail(), loginMember.getNickname());
-		loginMember.updateRefreshToken(tokenRequest.refreshToken());
-
-		return AuthenticationMapper.toLoginResponse(tokenRequest.accessToken(), tokenRequest.refreshToken());
+		return authorizationService
+			.issueServiceToken(httpServletResponse, loginMember.getEmail(), loginMember.getNickname());
 	}
 
 	public MemberInfoResponse searchMemberInfo(Long memberId) {
